Revert to default precision in the hessian_accumulate

+++ src/llmcompressor/modifiers/pruning/sparsegpt/sgpt_sparsify.py
@@ -5,7 +5,7 @@ import torch
 import transformers
 from loguru import logger
 
-SGPT_PRECISION = torch.float16
+SGPT_PRECISION = torch.float32
 
 
 def make_empty_hessian(
